Swap functions around to be at the right spots
